No time to waste

+++ rest/src/main/java/cz/idomatojde/rest/controllers/TimetableChatMessageController.java
@@ -4,16 +4,21 @@ import cz.idomatojde.dto.timetable.AddTimetableChatMessageDTO;
 import cz.idomatojde.dto.timetable.ChangeTextTimetableChatMessageDTO;
 import cz.idomatojde.dto.timetable.TimetableChatMessageDTO;
 import cz.idomatojde.facade.TimetableChatMessageFacade;
-import cz.idomatojde.rest.controllers.base.BaseRESTController;
+import cz.idomatojde.facade.UserFacade;
+import cz.idomatojde.rest.controllers.base.AuthBaseRESTController;
 import io.swagger.annotations.Api;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.inject.Inject;
 import java.util.List;
+
+import static org.springframework.http.ResponseEntity.ok;
 
 /**
  * Controller responsible for all things concerning TimetableChatMessages
@@ -23,29 +28,40 @@ import java.util.List;
 @Api(tags = "TimetableChatMessages Endpoint")
 @RestController
 @RequestMapping("timetableChatMessages")
-public class TimetableChatMessageController extends BaseRESTController<TimetableChatMessageFacade, AddTimetableChatMessageDTO, TimetableChatMessageDTO> {
+public class TimetableChatMessageController extends
+        AuthBaseRESTController<TimetableChatMessageFacade, AddTimetableChatMessageDTO, TimetableChatMessageDTO> {
     @Inject
-    public TimetableChatMessageController(TimetableChatMessageFacade chatMessages) {
-        super(chatMessages);
+    public TimetableChatMessageController(UserFacade userFacade, TimetableChatMessageFacade chatMessages) {
+        super(userFacade, chatMessages);
     }
 
     @GetMapping("allMessagesOfUser?userId={userId}")
-    List<TimetableChatMessageDTO> getMessagesByUserId(@PathVariable long userId){
-        return facade.getAllMessagesOfUser(userId);
+    ResponseEntity<List<TimetableChatMessageDTO>> getMessagesByUserId(@RequestHeader(value = "token") String token, @PathVariable long userId) {
+        if (notAuthenticated(token)) return unauthorized(null);
+
+        return ok(facade.getAllMessagesOfUser(userId));
     }
 
     @GetMapping("allMessagesOfTimetableEntry?entryId={entryId}")
-    List<TimetableChatMessageDTO> getAllMessagesOfTimetableEntry(@PathVariable long entryId){
-        return facade.getAllMessagesOfTimetableEntry(entryId);
+    ResponseEntity<List<TimetableChatMessageDTO>> getAllMessagesOfTimetableEntry(@RequestHeader(value = "token") String token, @PathVariable long entryId) {
+        if (notAuthenticated(token)) return unauthorized(null);
+
+        return ok(facade.getAllMessagesOfTimetableEntry(entryId));
     }
 
     @PostMapping("changeMessage")
-    void changeMessage(ChangeTextTimetableChatMessageDTO dto){
+    ResponseEntity<Void> changeMessage(@RequestHeader(value = "token") String token, ChangeTextTimetableChatMessageDTO dto) {
+        if (notAuthenticated(token)) return unauthorized();
+
         facade.changeText(dto);
+        return ok().build();
     }
 
     @PostMapping("deleteAllMessages?userId={userId}")
-    void deleteMessagesByUserId(@PathVariable long userId){
+    ResponseEntity<Void> deleteMessagesByUserId(@RequestHeader(value = "token") String token, @PathVariable long userId) {
+        if (notAuthenticated(token)) return unauthorized();
+
         facade.deleteAllMessagesOfUser(userId);
+        return ok().build();
     }
 }
